Fix build Errors in /bookings and /booking
