+++ apps/web/app/future/booking/[uid]/embed/page.tsx
@@ -1,9 +1,11 @@
-import OldPage, { type PageProps, getServerSideProps } from "@pages/booking/[uid]";
+import { type PageProps, getServerSideProps } from "@pages/booking/[uid]";
+import OldPage from "@pages/booking/[uid]/embed";
 import { withAppDirSsr } from "app/WithAppDirSsr";
 import withEmbedSsrAppDir from "app/WithEmbedSSR";
 import { WithLayout } from "app/layoutHOC";
 
 const getData = withAppDirSsr<PageProps>(getServerSideProps);
+
 const getEmbedData = withEmbedSsrAppDir(getData);
 
 export default WithLayout({ getLayout: null, getData: getEmbedData, Page: OldPage });
