+++ apps/web/lib/bookings/[status]/getStaticProps.tsx
@@ -1,0 +1,24 @@
+import { type GetStaticProps } from "next";
+import { z } from "zod";
+
+import { ssgInit } from "@server/lib/ssg";
+
+const validStatuses = ["upcoming", "recurring", "past", "cancelled", "unconfirmed"] as const;
+
+const querySchema = z.object({
+  status: z.enum(validStatuses),
+});
+
+export const getStaticProps: GetStaticProps = async (ctx) => {
+  const params = querySchema.safeParse(ctx.params);
+  const ssg = await ssgInit(ctx);
+
+  if (!params.success) return { notFound: true };
+
+  return {
+    props: {
+      status: params.data.status,
+      trpcState: ssg.dehydrate(),
+    },
+  };
+};
